Rename to flavours.
