+++ virtualenv/builders/base.py
@@ -19,7 +19,7 @@ SCRIPT_DIR = os.path.join(
 
 class BaseBuilder(object):
 
-    def __init__(self, python, system, system_site_packages=False, clear=False,
+    def __init__(self, python, flavour, system_site_packages=False, clear=False,
                  pip=True, setuptools=True, extra_search_dirs=None,
                  prompt=""):
         # We default to sys.executable if we're not given a Python.
@@ -31,7 +31,7 @@ class BaseBuilder(object):
             extra_search_dirs = []
 
         self.python = python
-        self.system = system
+        self.flavour = flavour
         self.system_site_packages = system_site_packages
         self.clear = clear
         self.pip = pip
@@ -69,7 +69,7 @@ class BaseBuilder(object):
 
     def install_scripts(self, destination):
         # Determine the list of files based on if we're running on Windows
-        files = set(self.system.activation_scripts)
+        files = self.flavour.activation_scripts
 
         # We just always want add the activate_this.py script regardless of
         # platform.
